[GH-74] Auto balance SP to create NAS server
The current implementation is that: if SP is not passed in (is None)
to create NAS server, storops will use the first item of SP list.

This feature is to auto balance the SP of creating NAS server when
SP is not passed in. SPB will be chosen only when the existing NAS
servers on SPA are more than the ones on SPB. Otherwise, SPA will
be chosen.