[GH-84] Error parsing VNX storage group output.
When the HBA uid is too long, the table format of storage group output
is messed up.  There is no space between each column.

Luckly, the format for storage processor is fixed.  We could use it as the
token.  Modify the regular expression to parse the HBA and verify the
parsing of storage group with the messed output.

Bump version to 0.4.4.