[GH-238] HLU number greater than 255
Unity supports maximum HLU number is 16381. So storops supports to
random pick a number from 0 to 16381 as the HLU when attaching the LUN
to host.

However, HLU number greater than 255 affects the operation of extending
an attached volume in OpenStack. The reason is that os-brick invokes
command `sg_scan` to get device information when extending volume, while
`sg_scan` uses 8 bits (maximum number is 255) to store the LUN id.
Any number greater than 255 will be truncated to wrong number via `and`
operation with `0xff`.

Thus, we need to update the range to [0 , 255].