[#305] VNX: alu-hlu cache dirty
This commit first reverts "Fix alu/hlu cache issue (#181)",
the commit id is e82db793b73c224f6b8b0f6aa222d6f666846cec.

The root cause of #181 was `sg.update()` would flush the alu-hlu
mapping which was not returned by `sg.update()` but the LUN/alu was
attached successfully on VNX. VNX was slow, which caused the info
returnd by `sg.update()` was not the latest one.

To fix the issue of #181, a retry is added to `detach_alu`. That is,
when it found the alu-hlu is not in the cache, it will update sg's
alu-hlu-map and try to get the hlu again.